breaking-api: Rename module texts to vars
- Some of them are not just texts
