+++ src/components/pages/LegPoseWidgets.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react"
 import { Card } from "../generic"
-import { ANGLE_NAMES, RANGE_PARAMS } from "../texts"
+import { ANGLE_NAMES, RANGE_PARAMS } from "../vars"
 
 class LegPoseWidget extends Component {
     updateFieldState = (angle, value) => {
